backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
oss2==2.18.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
